smart-invoice-extractor/requirements.txt
pdfplumber
pytesseract
pillow
pandas
openpyxl
streamlit
